Compile combination generators in emscripten, link to them in makefile
Also update emscripten code will call to optimize() template function

+++ include/table.h
@@ -1,7 +1,7 @@
 #include <boost/multi_array.hpp>
 
 typedef boost::multi_array<int, 2> table_type;
-typedef boost::multi_array_ref<int, 2> table_type_ref;
+typedef boost::multi_array_ref<uint16_t, 2> table_type_ref;
 // index type is not dependent on template parameters,
 // and is identical for multi_array and multi_array_ref classes
 typedef table_type::index table_index;
